LF-4695 Move Ensemble-related code from controller into Ensemble util (service)
