+++ packages/api/src/controllers/sensorController.js
@@ -30,21 +30,16 @@ import { transaction, Model } from 'objection';
 
 import {
   ENSEMBLE_BRAND,
-  getEnsembleOrganisations,
-  getOrganisationDevices,
-  calculateSensorArrayPoint,
-  enrichWithMockData,
   extractEsids,
   registerFarmAndClaimSensors,
   unclaimSensor,
   ENSEMBLE_UNITS_MAPPING,
-  ENSEMBLE_READING_TYPES_MAPPING,
+  getEnsembleSensors,
 } from '../util/ensemble.js';
 import { databaseUnit } from '../util/unit.js';
 import { sensorErrors, parseSensorCsv } from '../../../shared/validation/sensorCSV.js';
 import syncAsyncResponse from '../util/syncAsyncResponse.js';
 import knex from '../util/knex.js';
-import FarmModel from '../models/farmModel.js';
 
 const getSensorTranslations = async (language) => {
   // Remove country identifier from language preference
@@ -114,87 +109,8 @@ const sensorController = {
   },
   async getSensors(req, res) {
     const { farm_id } = req.headers;
-
     try {
-      const { id: EnsemblePartnerId } = await AddonPartnerModel.getPartnerId(ENSEMBLE_BRAND);
-
-      const { access_token } = await AddonPartnerModel.getAccessAndRefreshTokens(ENSEMBLE_BRAND);
-
-      const farmEnsembleAddon = await FarmAddonModel.getOrganisationId(farm_id, EnsemblePartnerId);
-
-      if (!farmEnsembleAddon) {
-        return res.status(200).send({ sensors: [], sensor_arrays: [] });
-      }
-
-      const farmEnsembleOrganisationid = farmEnsembleAddon.org_uuid;
-
-      // Will no longer be necessary once the primary key is stored on the farm_integration/farm_addon table
-      const allRegisteredOrganisations = await getEnsembleOrganisations(access_token);
-
-      const organisation = allRegisteredOrganisations.find(
-        ({ uuid }) => uuid === farmEnsembleOrganisationid,
-      );
-
-      const devices = await getOrganisationDevices(organisation.pk, access_token);
-
-      if (!devices.length) {
-        return res.status(200).send({ sensors: [], sensor_arrays: [] });
-      }
-
-      const sensors = [];
-      const sensorArrayMap = {};
-
-      // See below
-      const farm = await FarmModel.query().findById(farm_id);
-      const farmCenterCoordinates = farm.grid_points;
-
-      for (const incomingDevice of devices) {
-        if (incomingDevice.category === 'Sensor' && incomingDevice.deployed) {
-          // This is a temporary solution until the Ensemble API is updated to return depth, position, and profiles. Then please delete the enriching function and the FarmModel query
-          const device = enrichWithMockData(incomingDevice, farmCenterCoordinates);
-
-          const sensor = {
-            name: device.name,
-            external_id: device.esid,
-            sensor_reading_types: device.parameter_types.map(
-              (type) => ENSEMBLE_READING_TYPES_MAPPING[type],
-            ),
-            last_seen: device.last_seen,
-            point: {
-              lat: device.latest_position.coordinates.lat,
-              lng: device.latest_position.coordinates.lng,
-            },
-            depth: device.latest_position.depth,
-            depth_unit: 'cm', // to be confirmed
-            sensor_array_id: device.profile_id,
-
-            // The following only for backwards compatibility until old sensor flow is removed
-            location_id: device.esid,
-          };
-          sensors.push(sensor);
-
-          if (device.profile_id) {
-            if (!sensorArrayMap[device.profile_id]) {
-              sensorArrayMap[device.profile_id] = [];
-            }
-            sensorArrayMap[device.profile_id].push({
-              external_id: sensor.external_id,
-              // used to calculate sensor_array.point
-              latest_position: device.latest_position,
-            });
-          }
-        }
-      }
-
-      const sensor_arrays = Object.entries(sensorArrayMap).map(([id, sensors]) => ({
-        id,
-        sensors: sensors.map(({ external_id }) => external_id),
-        point: calculateSensorArrayPoint(sensors),
-
-        // The following only for backwards compatibility until old sensor flow is removed
-        location_id: id,
-        name: `Sensor Array ${id}`,
-      }));
+      const { sensors, sensor_arrays } = await getEnsembleSensors(farm_id);
 
       return res.status(200).send({
         sensors,
